Extract `getSubmittableQuestion`
